requirements.txt
Flask
spotipy
python-dotenv 